backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
pyarrow==15.0.0
